refactor(platform): split log-detail-page into separate component files
- extract components into log-detail/components/ folder
  - agent-events-card.tsx
  - artifact-download-button.tsx
  - copy-field.tsx
  - event-type-filter-dropdown.tsx
  - formatted-events-view.tsx
  - info-row.tsx
  - log-detail-content.tsx
  - raw-json-view.tsx
  - view-mode-toggle.tsx
- extract utility functions into log-detail/utils.ts
- maintain backward compatibility via re-export
