+++ turbo/apps/platform/src/views/logs-page/log-detail/index.tsx
@@ -1,0 +1,27 @@
+import { useGet } from "ccstate-react";
+import { AppShell } from "../../layout/app-shell.tsx";
+import { currentLogId$ } from "../../../signals/logs-page/log-detail-state.ts";
+import { LogDetailContent } from "./components/log-detail-content.tsx";
+
+export function LogDetailPage() {
+  const logId = useGet(currentLogId$);
+
+  const breadcrumb = [
+    { label: "Logs", path: "/logs" as const },
+    { label: logId ? `Run ID - ${logId}` : "Detail" },
+  ];
+
+  return (
+    <AppShell breadcrumb={breadcrumb}>
+      <div className="px-6 py-4">
+        {logId ? (
+          <LogDetailContent logId={logId} />
+        ) : (
+          <div className="p-8 text-center text-muted-foreground">
+            Log ID not found
+          </div>
+        )}
+      </div>
+    </AppShell>
+  );
+}
